requirements.txt
# Heart Disease Classification Project - Requirements
# EAS 503 Final Project

# Core ML Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Hyperparameter Tuning
optuna>=3.5.0

# Experiment Tracking
mlflow>=2.10.0
python-dotenv>=1.0.0

# Data Download
kagglehub>=0.2.0

# Jupyter Notebooks
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# API (FastAPI)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Frontend (Streamlit)
streamlit>=1.30.0
requests>=2.31.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter for Render

# Visualization (optional, for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0